notice delete, insert admin

+++ src/main/java/com/farmorai/backend/controller/NoticeController.java
@@ -19,6 +19,8 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Map;
 
+import static org.springframework.http.HttpStatus.*;
+
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/notice")
@@ -42,6 +44,9 @@ public class NoticeController {
                                            @RequestBody NoticeDto noticeDto) {
         log.info(userDetails);
         log.info(SecurityContextHolder.getContext().getAuthentication());
+        if(userDetails == null || !userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ROLE_ADMIN"))){
+            return ResponseEntity.status(FORBIDDEN).body("관리자만 글 등록이 가능합니다.");
+        }
         noticeService.insertNotice(userDetails.getMemberId(),noticeDto);
         return ResponseEntity.ok("글 등록 성공");
     }
@@ -54,8 +59,12 @@ public class NoticeController {
     }
 
     @DeleteMapping("/{noticeId}")
-    public Map<String,String> deleteNotice(@PathVariable Long noticeId) {
+    public ResponseEntity<?> deleteNotice(@AuthenticationPrincipal CustomUserDetails userDetails,
+                                           @PathVariable Long noticeId) {
+        if(userDetails == null || !userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ROLE_ADMIN"))){
+            return ResponseEntity.status(FORBIDDEN).body("관리자만 공지사항을 삭제할 수 있습니다.");
+        }
         noticeService.deleteNotice(noticeId);
-        return Map.of("result","success");
+        return ResponseEntity.ok("공지사항이 삭제되었습니다.");
     }
 }
